Code/backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.2
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.0
bleach==6.0.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.402
charset-normalizer==3.2.0
checksumdir==1.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
executing==1.2.0
fastapi==0.99.1
fastapi-socketio==0.0.10
fastjsonschema==2.18.0
filelock==3.12.3
filetype==1.2.0
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
googleapis-common-protos==1.60.0
greenlet==3.0.3
grpcio==1.58.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kombu==5.3.4
langchain==0.0.287
langsmith==0.0.36
Lazify==0.4.0
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
nodeenv==1.8.0
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.25.0
onnxruntime==1.15.1
openai==0.28.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
posthog==3.0.2
prisma==0.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pypdf==3.16.0
PyPika==0.48.9
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
Send2Trash==1.8.2
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.14.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uptrace==1.20.0
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
vine==5.1.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.0.1
Whoosh==2.7.4
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
zipp==3.16.2
